Fix minor compiler warning.

+++ src/emucore/CartBUS.cxx
@@ -71,7 +71,8 @@ CartridgeBUS::CartridgeBUS(const ByteBuffer& image, size_t size,
     0x00000808,
     0x40001FDC,
     devSettings ? settings.getBool("dev.thumb.trapfatal") : false,
-    devSettings ? settings.getFloat("dev.thumb.cyclefactor") : 1.0,
+    devSettings ? static_cast<double>(
+        settings.getFloat("dev.thumb.cyclefactor")) : 1.0,
     Thumbulator::ConfigureFor::BUS,
     this);
